Editor styled. Utils moved to their own file.
